cast results to int to give round numbers.

+++ munin/mod/maxcap.py
@@ -68,8 +68,8 @@ class maxcap(loadable.loadable):
             cap_rate=u.planet.cap_rate(victim)
         for i in range(1,5):
             cap+=int(total_roids*cap_rate)
-            reply+="Wave %d: %d (%d), " % (i,total_roids*cap_rate,cap)
-            total_roids = total_roids - total_roids*cap_rate
+            reply+="Wave %d: %d (%d), " % (i,int(total_roids*cap_rate),cap)
+            total_roids = total_roids - int(total_roids*cap_rate)
         irc_msg.reply("Caprate: %s%% %s"%(cap_rate*100,reply.strip(', '))
 
         return 1
